+++ src/App.tsx
@@ -1,276 +1,29 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { useEffect, useState } from "react";
-import { Web3Auth } from "@web3auth/modal";
-import {
-  CHAIN_NAMESPACES,
-  IProvider,
-  WEB3AUTH_NETWORK,
-  WALLET_ADAPTERS,
-} from "@web3auth/base";
-import Web3 from "web3";
-
 import "./App.css";
-import { EthereumPrivateKeyProvider } from "@web3auth/ethereum-provider";
 
-import { MetamaskAdapter } from "@web3auth/metamask-adapter";
-
-// IMP START - Dashboard Registration
-const clientId =
-  "BPbO-LorL6VnxrYGqX9WrY23EIN1cEEz9qR1Ir4npgxR8Yik9WfXh8_ic8o7en7yN7usdzHNYb8fEQxokEUzI_E"; // get from https://dashboard.web3auth.io
-
-const chainConfig = {
-  chainId: "0x13882", // Please use 0x1 for Mainnet
-  rpcTarget: "https://rpc-amoy.polygon.technology",
-  chainNamespace: CHAIN_NAMESPACES.EIP155,
-  displayName: "Amoy",
-  blockExplorerUrl: "https://www.oklink.com/amoy",
-  ticker: "MATIC",
-  tickerName: "MATIC",
-};
-
-const privateKeyProvider = new EthereumPrivateKeyProvider({
-  config: { chainConfig: chainConfig },
-});
-
-const web3auth = new Web3Auth({
-  uiConfig: {
-    defaultLanguage: "ko",
-    loginGridCol: 2,
-    loginMethodsOrder: ["google", "kakao", "github"],
-    primaryButton: "emailLogin",
-  },
-  clientId,
-  web3AuthNetwork: WEB3AUTH_NETWORK.SAPPHIRE_DEVNET,
-  privateKeyProvider: privateKeyProvider,
-});
-
-web3auth.configureAdapter(new MetamaskAdapter());
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Home from "./pages/Home";
+import Login from "./pages/Login";
+import Setting from "./pages/Setting";
+import Faq from "./pages/Faq";
+import Qna from "./pages/Qna";
+import Mypage from "./pages/Mypage";
+import Profile from "./pages/Profile";
 
 function App() {
-  const [provider, setProvider] = useState<IProvider | null>(null);
-  const [loggedIn, setLoggedIn] = useState(false);
-
-  useEffect(() => {
-    const init = async () => {
-      try {
-        await web3auth.initModal({
-          modalConfig: {
-            [WALLET_ADAPTERS.OPENLOGIN]: {
-              label: "openlogin",
-              loginMethods: {
-                github: {
-                  name: "github",
-                  showOnModal: false,
-                },
-                reddit: {
-                  name: "reddit",
-                  showOnModal: false,
-                },
-                discord: {
-                  name: "discord",
-                  showOnModal: false,
-                },
-                twitch: {
-                  name: "twitch",
-                  showOnModal: false,
-                },
-                line: {
-                  name: "line",
-                  showOnModal: false,
-                },
-                linkedin: {
-                  name: "linkedin",
-                  showOnModal: false,
-                },
-                weibo: {
-                  name: "weibo",
-                  showOnModal: false,
-                },
-                wechat: {
-                  name: "wechat",
-                  showOnModal: false,
-                },
-                farcaster: {
-                  name: "farcaster",
-                  showOnModal: false,
-                },
-              },
-            },
-          },
-        });
-        setProvider(web3auth.provider);
-
-        if (web3auth.connected) {
-          setLoggedIn(true);
-        }
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    init();
-  }, []);
-
-  const login = async () => {
-    const web3authProvider = await web3auth.connect();
-    console.log();
-    setProvider(web3authProvider);
-    if (web3auth.connected) {
-      setLoggedIn(true);
-    }
-  };
-
-  const authenticateUser = async () => {
-    if (!web3auth) {
-      uiConsole("web3auth not initialized yet");
-      return;
-    }
-    const idToken = await web3auth.authenticateUser();
-    uiConsole(idToken);
-  };
-
-  const getUserInfo = async () => {
-    const user = await web3auth.getUserInfo();
-    uiConsole(user);
-  };
-
-  const logout = async () => {
-    await web3auth.logout();
-    setProvider(null);
-    setLoggedIn(false);
-    uiConsole("logged out");
-  };
-
-  const getAccounts = async () => {
-    if (!provider) {
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    const web3 = new Web3(provider as any);
-
-    // Get user's Ethereum public address
-    const address = await web3.eth.getAccounts();
-    uiConsole(address);
-  };
-
-  const getBalance = async () => {
-    if (!provider) {
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    const web3 = new Web3(provider as any);
-
-    // Get user's Ethereum public address
-    const address = (await web3.eth.getAccounts())[0];
-
-    // Get user's balance in ether
-    const balance = web3.utils.fromWei(
-      await web3.eth.getBalance(address), // Balance is in wei
-      "ether"
-    );
-    uiConsole(balance);
-  };
-
-  const signMessage = async () => {
-    if (!provider) {
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    const web3 = new Web3(provider as any);
-
-    // Get user's Ethereum public address
-    const fromAddress = (await web3.eth.getAccounts())[0];
-
-    const originalMessage = "YOUR_MESSAGE";
-
-    // Sign the message
-    const signedMessage = await web3.eth.personal.sign(
-      originalMessage,
-      fromAddress,
-      "test password!" // configure your own password here.
-    );
-    uiConsole(signedMessage);
-  };
-
-  function uiConsole(...args: any[]): void {
-    const el = document.querySelector("#console>p");
-    if (el) {
-      el.innerHTML = JSON.stringify(args || {}, null, 2);
-    }
-    console.log(...args);
-  }
-
-  const loggedInView = (
-    <>
-      <div className="flex-container">
-        <div>
-          <button onClick={getUserInfo} className="card">
-            Get User Info
-          </button>
-        </div>
-        <div>
-          <button onClick={authenticateUser} className="card">
-            Get ID Token
-          </button>
-        </div>
-        <div>
-          <button onClick={getAccounts} className="card">
-            Get Accounts
-          </button>
-        </div>
-        <div>
-          <button onClick={getBalance} className="card">
-            Get Balance
-          </button>
-        </div>
-        <div>
-          <button onClick={signMessage} className="card">
-            Sign Message
-          </button>
-        </div>
-        <div>
-          <button onClick={logout} className="card">
-            Log Out
-          </button>
-        </div>
-      </div>
-    </>
-  );
-
-  const unloggedInView = (
-    <button onClick={login} className="card">
-      Login
-    </button>
-  );
-
   return (
-    <div className="container">
-      <h1 className="title">
-        <a
-          target="_blank"
-          href="https://web3auth.io/docs/sdk/pnp/web/modal"
-          rel="noreferrer"
-        >
-          Web3Auth{" "}
-        </a>
-        & ReactJS (Webpack) Quick Start
-      </h1>
-
-      <div className="grid">{loggedIn ? loggedInView : unloggedInView}</div>
-      <div id="console" style={{ whiteSpace: "pre-line" }}>
-        <p style={{ whiteSpace: "pre-line" }}></p>
+    <BrowserRouter>
+      <div className="App">
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/mypage" element={<Mypage />} />
+          <Route path="/profile" element={<Profile />} />
+          <Route path="/setting" element={<Setting />} />
+          <Route path="/faq" element={<Faq />} />
+          <Route path="/qna" element={<Qna />} />
+        </Routes>
       </div>
-
-      <footer className="footer">
-        <a
-          href="https://github.com/Web3Auth/web3auth-pnp-examples/tree/main/web-modal-sdk/quick-starts/react-modal-quick-start"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Source code
-        </a>
-      </footer>
-    </div>
+    </BrowserRouter>
   );
 }
